% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{calc_tb}
\alias{calc_tb}
\title{calc_tb}
\usage{
calc_tb(h, doubles, triples, hr)
}
\arguments{
\item{h}{total hits}

\item{doubles}{num doubles}

\item{triples}{num triples}

\item{hr}{num homers}
}
\value{
numeric vector
}
\description{
calc_tb
}

