% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{calculate_prices}
\alias{calculate_prices}
\title{calculate prices}
\usage{
calculate_prices(pp_list, hit_pitch)
}
\arguments{
\item{pp_list}{named list of zscored stats. look at `proj_prep` to see
where this function gets called}

\item{hit_pitch}{c('h', 'p')}
}
\value{
data frame, with `value` variable.
}
\description{
takes zscores over replacement and converts to auction values
}

