% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_pos}
\alias{clean_pos}
\title{clean pos}
\usage{
clean_pos(x)
}
\arguments{
\item{x}{vector of positions}
}
\value{
cleaned character vector
}
\description{
converts invalid position types to valid
}

