% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_quotes}
\alias{clean_quotes}
\title{clean quotes from a string}
\usage{
clean_quotes(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector w/o quotes
}
\description{
clean quotes from a string
}

