% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangraphs.R
\name{clean_raw_fangraphs}
\alias{clean_raw_fangraphs}
\title{Cleans up a fangraphs (steamer, zips etc) projection file.}
\usage{
clean_raw_fangraphs(df, hit_pitch)
}
\arguments{
\item{df}{raw fangraphs df, eg output of read_raw_steamer}

\item{hit_pitch}{c('h', 'p')}
}
\value{
a data frame with consistent variable names
}
\description{
names, consistent stat names, etc.
}

