% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy_pros.R
\name{clean_raw_fantasy_pros}
\alias{clean_raw_fantasy_pros}
\title{Cleans up a fantasy pros projection scrape}
\usage{
clean_raw_fantasy_pros(df, hit_pitch)
}
\arguments{
\item{df}{raw fantasy pros df.  output of read_raw_fantasy_pros.}

\item{hit_pitch}{c('h', 'p')}
}
\value{
a data frame with consistent variable names
}
\description{
names, consistent stat names, etc.
}

