% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pod.R
\name{clean_raw_pod}
\alias{clean_raw_pod}
\title{Cleans up a pod projection file.}
\usage{
clean_raw_pod(df, hit_pitch)
}
\arguments{
\item{df}{raw pod df.  output of read_raw_pod}

\item{hit_pitch}{c('h', 'p')}
}
\value{
a data frame with consistent variable names
}
\description{
names, consistent stat names, etc.
}

