% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamer.R
\name{clean_raw_steamer}
\alias{clean_raw_steamer}
\title{Cleans up a steamer projection file.}
\usage{
clean_raw_steamer(df, hit_pitch)
}
\arguments{
\item{df}{raw steamer df.  output of read_raw_steamer.}

\item{hit_pitch}{c('h', 'p')}
}
\value{
a data frame with consistent variable names
}
\description{
names, consistent stat names, etc.
}

