% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_prep.R
\name{common_proj_prep_vars}
\alias{common_proj_prep_vars}
\title{What common variables do all projection prep data frames need to have?}
\usage{
common_proj_prep_vars()
}
\value{
vector with names of common variables
}
\description{
What common variables do all projection prep data frames need to have?
}

