% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.R
\name{find_mlbid}
\alias{find_mlbid}
\title{find_mlbid}
\usage{
find_mlbid(selected_player, match_df)
}
\arguments{
\item{selected_player}{display name of the player to look up}

\item{match_df}{slim data frame to look in.  should include disp_name
and mlbid.}
}
\value{
mlbid of matching player
}
\description{
find_mlbid
}

