% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{find_special_replacement}
\alias{find_special_replacement}
\title{find special replacement}
\usage{
find_special_replacement(pp_list, hit_pitch)
}
\arguments{
\item{pp_list}{named list of zscored stats. look at `proj_prep` to see
where this function gets called}

\item{hit_pitch}{c('h', 'p')}
}
\value{
named list, with replacement player for special positions and
data frame with mlbid / position for determining replacement
}
\description{
finds replacement level (and replacement position) for
non-standard (eg util) positions
}

