% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{find_standard_replacement}
\alias{find_standard_replacement}
\title{Finds the mlbid of the replacement player, for each standard position}
\usage{
find_standard_replacement(pp_list, hit_pitch)
}
\arguments{
\item{pp_list}{named list of zscored stats. look at `proj_prep` to see
where this function gets called}

\item{hit_pitch}{c('h', 'p')}
}
\value{
named list, with mlbids of replacement player per-position
}
\description{
Finds the mlbid of the replacement player, for each standard position
}

