% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbid_matching.R
\name{first_pass_mlbid}
\alias{first_pass_mlbid}
\title{first pass mlbid match}
\usage{
first_pass_mlbid(name_vector)
}
\arguments{
\item{name_vector}{character vector of names to match}
}
\value{
vector of mlbids
}
\description{
first get the easy ones - the players for whom there
is one, and only one, match on the exact spelling of their full name
}

