% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{force_numeric}
\alias{force_numeric}
\title{force numeric}
\usage{
force_numeric(df, cols)
}
\arguments{
\item{df}{data frame}

\item{cols}{vector of columns that match names in the df}
}
\value{
data frame
}
\description{
given a df, and a vector of columns, force those variables
from character to numeric
}

