% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs.R
\name{get_cbs}
\alias{get_cbs}
\title{Get cbs}
\usage{
get_cbs(year)
}
\arguments{
\item{year}{}
}
\value{
projection prep object
}
\description{
workhorse function.  reads the raw cbs data file,
cleans up headers, returns projection prep object
}

