% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangraphs.R
\name{get_fangraphs}
\alias{get_fangraphs}
\title{Get various fangraphs projections}
\usage{
get_fangraphs(year, proj_system, limit_unmatched = TRUE)
}
\arguments{
\item{year}{desired year.  valid values: 2016}

\item{proj_system}{proj_system name of a projection system.  one of c('zips',
'fan', 'steamer', 'steamer600', 'fangraphsdc')}

\item{limit_unmatched}{if TRUE (the default behavior) will only
return players with an mlbid that can be matched.  look at `id_map`
and the `universal_metadata` vignette for more about the id map
we're using to match players to ids.
fundamentally, you need a consistent, unique identifier if you
want to work with multiple projection systems.  so this really
needs to be TRUE.}
}
\value{
list of named projection data frames.
}
\description{
workhorse function.  reads, cleans, preps, matches
fangraphs projections
}

