% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guru.R
\name{get_guru}
\alias{get_guru}
\title{Get guru projections}
\usage{
get_guru(year)
}
\arguments{
\item{year}{desired year.  valid values: 2016}
}
\value{
list of named projection data frames.
}
\description{
workhorse function.  reads the raw guru excel file,
cleans up headers, returns list of projection data frames ready for
projection_prep function.
}

