% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamer.R
\name{get_steamer}
\alias{get_steamer}
\title{Get steamer projections}
\usage{
get_steamer(year)
}
\arguments{
\item{year}{desired year.  valid values: 2016}
}
\value{
a projection prep object.
}
\description{
workhorse function.  reads the raw steamer data file,
cleans up headers, returns consistent df
}

