% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{id_map}
\alias{id_map}
\title{id_map}
\format{\describe{
\item{idplayer}{idplayer}
\item{playername}{playername}
\item{birthdate}{birthdate}
\item{firstname}{firstname}
\item{lastname}{lastname}
\item{team}{team}
\item{lg}{lg}
\item{pos}{pos}
\item{idfangraphs}{idfangraphs}
\item{fangraphsname}{fangraphsname}
\item{mlbid}{mlbid}
\item{mlbname}{mlbname}
\item{cbsid}{cbsid}
\item{cbsname}{cbsname}
\item{retroid}{retroid}
\item{brefid}{brefid}
\item{nfbcid}{nfbcid}
\item{nfbcname}{nfbcname}
\item{espnid}{espnid}
\item{espnname}{espnname}
\item{kfflname}{kfflname}
\item{davenportid}{davenportid}
\item{bpid}{bpid}
\item{yahooid}{yahooid}
\item{yahooname}{yahooname}
\item{mstrbllname}{mstrbllname}
\item{bats}{bats}
\item{throws}{throws}
\item{fantprosname}{fantprosname}
\item{lastcommafirst}{lastcommafirst}
\item{rotowireid}{rotowireid}
\item{fanduelname}{fanduelname}
\item{fanduelid}{fanduelid}
\item{draftkingsname}{draftkingsname}
\item{ottoneuid}{ottoneuid}
\item{hqid}{hqid}
}}
\source{
smartfantasybaseball.com
}
\usage{
id_map
}
\description{
dataframe that maps names/ids across various mlb data sources, by
Tanner: http://www.smartfantasybaseball.com/tools/
}
\keyword{datasets}

