% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_prep.R
\name{limit_proj_vars}
\alias{limit_proj_vars}
\title{select relevant raw projection stats}
\usage{
limit_proj_vars(proj_list)
}
\arguments{
\item{proj_list}{named ('h', 'p') list of data frames, ie output of get_steamer}
}
\value{
list of data frames with relevant variables selected
}
\description{
select relevant raw projection stats
}

