% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbid_matching.R
\name{mlbid_match}
\alias{mlbid_match}
\title{mlbid match}
\usage{
mlbid_match(df)
}
\arguments{
\item{df}{cleaned df, eg output of clean_raw_razzball_steamer}
}
\value{
vector of ids
}
\description{
wrapper around id matching functions that
attempts to match player names to an id
}

