% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.R
\name{prep_name_metadata}
\alias{prep_name_metadata}
\title{prep name metadata for optimal auction app}
\usage{
prep_name_metadata(id_map)
}
\arguments{
\item{id_map}{an id map with fields playername, pos, and mlbid.
presumably the same id map that is used for `mlbid_match`?}
}
\value{
slim data frame with mlbid, display_name, hit_pitch
}
\description{
prep name metadata for optimal auction app
}

