% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_prep.R
\name{proj_prep}
\alias{proj_prep}
\title{Create a proj_prep object}
\usage{
proj_prep(raw_proj, ...)
}
\arguments{
\item{raw_proj}{raw projection data.}
}
\description{
given raw projection data, creates a proj_prep object
}
\examples{
\dontrun{
cdf_mv <- mapvizieR(ex_CombinedAssessmentResults,
                    ex_CombinedStudentsBySchool)

is.mapvizieR(cdf_mv)
}
}

