% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamer.R
\name{read_and_clean_steamer}
\alias{read_and_clean_steamer}
\title{Reads and cleans steamer projection}
\usage{
read_and_clean_steamer(year)
}
\arguments{
\item{year}{desired year.  valid values: 2016}
}
\value{
data frame
}
\description{
workhorse function.  reads the raw steamer data file,
cleans up headers, returns consistent df
}

