% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs.R
\name{read_raw_cbs}
\alias{read_raw_cbs}
\title{Read raw cbs projections for a given year}
\usage{
read_raw_cbs(year)
}
\arguments{
\item{year}{desired year.  valid values: 2016}
}
\value{
named list of data frames
}
\description{
Read raw cbs projections for a given year
}

