% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy_pros.R
\name{read_raw_fantasy_pros}
\alias{read_raw_fantasy_pros}
\title{Read raw fantasy pros projections for a given year}
\usage{
read_raw_fantasy_pros(year)
}
\arguments{
\item{year}{desired year.  valid values: 2016}
}
\value{
named list of data frames
}
\description{
Read raw fantasy pros projections for a given year
}

