% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guru.R
\name{read_raw_guru}
\alias{read_raw_guru}
\title{Read raw guru projections for a given year}
\usage{
read_raw_guru(year)
}
\arguments{
\item{year}{desired year.  valid values: 2016}
}
\value{
named list of data frames
}
\description{
Read raw guru projections for a given year
}

