% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecota.R
\name{read_raw_pecota}
\alias{read_raw_pecota}
\title{Read raw pecota projections, given a path to the download file}
\usage{
read_raw_pecota(path_to_file)
}
\arguments{
\item{path_to_file}{local path to where the pod projections file is located.}
}
\value{
named list of data frames
}
\description{
reads the BP PECOTA projections.
these are paid projections, so this function requires
a local path as the argument.
}

