% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pod.R
\name{read_raw_pod}
\alias{read_raw_pod}
\title{Read raw guru projections, given a path to the download file}
\usage{
read_raw_pod(path_to_file)
}
\arguments{
\item{path_to_file}{local path to where the pod projections file is located.}
}
\value{
named list of data frames
}
\description{
reads the http://www.projectingx.com/baseball-player-projections/
pod projections.  pod projections are paid, so this function requires
a local path as the argument.
}

