% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/razz_steamer.R
\name{read_raw_razzball_steamer}
\alias{read_raw_razzball_steamer}
\title{Read raw steamer projections in for a given year}
\usage{
read_raw_razzball_steamer(year)
}
\arguments{
\item{year}{desired year.  valid values: 2016}
}
\value{
named list of data frames
}
\description{
this function will handle any logic for changing urls, etc.
the goal is to create a consistentcy in the R calls, so that whatever
work that needs to get done in locating data year over year is handled
by the functions and not exposed to the end user.
}

