% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{replacement_to_df}
\alias{replacement_to_df}
\title{replacement_to_df}
\usage{
replacement_to_df(pp_list, hit_pitch, special = FALSE)
}
\arguments{
\item{pp_list}{named list of zscored stats. look at `proj_prep` to see
where this function gets called}

\item{hit_pitch}{c('h', 'p')}

\item{special}{regular positions (c, 1b, 2b) (FALSE) or
special positions (TRUE)}
}
\value{
data frame
}
\description{
turns the replacement list into a data frame to assist
in calculating replacement level
}

