% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs.R
\name{scrape_cbs_h}
\alias{scrape_cbs_h}
\title{Scrape CBS Projeections}
\usage{
scrape_cbs_h(url, pos)
}
\arguments{
\item{url}{html table with steamer projections}

\item{pos}{position name}
}
\value{
data frame with fantasy pros projection data
}
\description{
Scrape CBS Projeections
}

