% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangraphs.R
\name{scrape_fangraphs}
\alias{scrape_fangraphs}
\title{generic fangraphs scrape}
\usage{
scrape_fangraphs(bat_pitch, proj_system)
}
\arguments{
\item{bat_pitch}{either 'bat' or 'pit'}

\item{proj_system}{name of a projection system.  one of c('zips',
'steamer', 'steamer600', 'fangraphsdc')}
}
\value{
data frame
}
\description{
pulls down fangraphs projections, given a projection
system and hit/pitch value
}

