% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy_pros.R
\name{scrape_fantasy_pros}
\alias{scrape_fantasy_pros}
\title{Scrape Fantasy Pros Projeections}
\usage{
scrape_fantasy_pros(url)
}
\arguments{
\item{url}{html table with steamer projections}
}
\value{
data frame with fantasy pros projection data
}
\description{
Scrape Fantasy Pros Projeections
}

