% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbid_matching.R
\name{second_pass_mlbid}
\alias{second_pass_mlbid}
\title{second pass id match}
\usage{
second_pass_mlbid(name_vector, mlbid_vector)
}
\arguments{
\item{name_vector}{vector of names}

\item{mlbid_vector}{vector of ids}
}
\value{
vector of ids, hopefully with more matches
}
\description{
second pass id match
}

