% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{special_positions_map}
\alias{special_positions_map}
\title{special_positions_map}
\format{\describe{
\item{CI}{corner infield}
\item{MI}{middle infield}
\item{Util}{utility}
\item{P}{pitcher}
}}
\usage{
special_positions_map
}
\description{
named list of 'special positions' (eg CI) with the corresponding
regular positions (eg 1B, 3B)
}
\keyword{datasets}

