% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{split_firstlast}
\alias{split_firstlast}
\title{Splits firstlast player name into first, last}
\usage{
split_firstlast(x)
}
\arguments{
\item{x}{player name, firstlast format ('Barack Obama')}
}
\value{
named list, with first and last name
}
\description{
Splits firstlast player name into first, last
}

