% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_defaults.R
\name{stock_defaults}
\alias{stock_defaults}
\title{stock defaults}
\usage{
stock_defaults()
}
\value{
'OK' if settings were saved.
}
\description{
the out-of-the-box stat/league/auction defaults.
unsurprisingly, these currently match the league settings for
almartin82's league.
}

