% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{tag_position}
\alias{tag_position}
\title{tag position priority}
\usage{
tag_position(player_positions, priorities)
}
\arguments{
\item{player_positions}{vector of positions (1 player, not vector of vectors)}

\item{priorities}{ranked heirarchy, usually from `user_settings`.}
}
\value{
vector of numeric priorities
}
\description{
tag position priority
}

