% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{value_over_replacement}
\alias{value_over_replacement}
\title{calculates value over replacement player, by position}
\usage{
value_over_replacement(pp_list, hit_pitch)
}
\value{
data frame, with zscore_replacement value added
}
\description{
calculates value over replacement player, by position
}

