% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy_pros.R
\name{get_fantasy_pros}
\alias{get_fantasy_pros}
\title{Get fantasy pros}
\usage{
get_fantasy_pros(year)
}
\arguments{
\item{year}{}
}
\value{
projection prep object
}
\description{
workhorse function.  reads the raw fantasy pros data file,
cleans up headers, returns projection prep object
}

