% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy_pros.R
\name{get_fantasy_pros}
\alias{get_fantasy_pros}
\title{Get fantasy pros}
\usage{
get_fantasy_pros(year, limit_unmatched = TRUE)
}
\arguments{
\item{year}{desired year.  valid values: 2016}

\item{limit_unmatched}{if TRUE (the default behavior) will only
return players with an mlbid that can be matched.  look at `id_map`
and the `universal_metadata` vignette for more about the id map
we're using to match players to ids.
fundamentally, you need a consistent, unique identifier if you
want to work with multiple projection systems.  so this really
needs to be TRUE.}
}
\value{
projection prep object
}
\description{
workhorse function.  reads the raw fantasy pros dat,
cleans up headers, returns list of projection data frames ready for
projection_prep function.
}

