% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pod.R
\name{get_pod}
\alias{get_pod}
\title{Get pod projections}
\usage{
get_pod(path_to_file)
}
\arguments{
\item{path_to_file}{path to the pod excel file you downloaded
(pod projections are paid).}
}
\value{
list of named projection data frames.
}
\description{
workhorse function.  reads the raw pod excel file,
cleans up headers, returns list of projection data frames ready for
projection_prep function.
}

