% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/razzball_steamer.R
\name{get_razzball_steamer}
\alias{get_razzball_steamer}
\title{Get steamer projections}
\usage{
get_razzball_steamer(year)
}
\arguments{
\item{year}{desired year.  valid values: 2016}
}
\value{
list of named projection data frames.
}
\description{
workhorse function.  reads the raw steamer data,
cleans up headers, returns list of projection data frames ready for
projection_prep function.
}

