% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamer.R
\name{scrape_steamer}
\alias{scrape_steamer}
\title{Scrape Steamer Projections}
\usage{
scrape_steamer(url)
}
\arguments{
\item{url}{html table with steamer projections}
}
\value{
data frame with steamer projection data
}
\description{
Scrape Steamer Projections
}

