% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{value_over_replacement}
\alias{value_over_replacement}
\title{calculates value over replacement player, by position}
\usage{
value_over_replacement(pp_list, hit_pitch)
}
\arguments{
\item{pp_list}{named list of zscored stats. look at `proj_prep` to see
where this function gets called}

\item{hit_pitch}{c('h', 'p')}
}
\value{
data frame, with zscore_replacement value added
}
\description{
calculates value over replacement player, by position
}

