% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{zscore}
\usage{
zscore(proj_list, hit_pitch, limit_player_pool = TRUE)
}
\arguments{
\item{proj_list}{named ('h', 'p') list of data frames, ie output of get_steamer}

\item{hit_pitch}{'h' or 'p'}

\item{limit_player_pool}{zscores relative to all players, or relative size of drafted
player pool?}
}
\value{
list of data frames with stat zscores
}
\description{
zscore
}

